Create Whirlpool_Lake.rms
Add whirlpool lake, a map featuring a whirlpool with plenty of fish in the center of the map. Players start in mangrove shallows that is connected to the lake through more shallows. The land surrounding the lake has lots of small depressions with cracked terrain and small stone deposits. 4 neutral markets are in the map's corners